moves another test from sli_retrieval_test.go to common_test.go
